Move demographics confirmation API call from confirmation page to apoointments list page
