+++ src/applications/check-in/day-of/pages/CheckIn/index.jsx
@@ -1,12 +1,14 @@
-import React, { useCallback, useMemo } from 'react';
+import React, { useCallback, useEffect, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { api } from '../../../api';
 import { useFormRouting } from '../../../hooks/useFormRouting';
 import { receivedMultipleAppointmentDetails } from '../../../actions/day-of';
-
+import { makeSelectFeatureToggles } from '../../../utils/selectors/feature-toggles';
 import DisplayMultipleAppointments from './DisplayMultipleAppointments';
+import { useSessionStorage } from '../../../hooks/useSessionStorage';
+import { useDemographicsFlags } from '../../../hooks/useDemographicsFlags';
 
 import {
   makeSelectVeteranData,
@@ -17,6 +19,16 @@ const CheckIn = props => {
   const { router } = props;
   const { t } = useTranslation();
   const { goToErrorPage } = useFormRouting(router);
+  const selectFeatureToggles = useMemo(makeSelectFeatureToggles, []);
+  const featureToggles = useSelector(selectFeatureToggles);
+  const { isDayOfDemographicsFlagsEnabled } = featureToggles;
+  const { getShouldSendDemographicsFlags } = useSessionStorage(false);
+  const {
+    demographicsData,
+    demographicsFlagsSent,
+    setDemographicsFlagsSent,
+    demographicsFlagsEmpty,
+  } = useDemographicsFlags();
   const selectVeteranData = useMemo(makeSelectVeteranData, []);
   const { appointments } = useSelector(selectVeteranData);
   const selectCurrentContext = useMemo(makeSelectCurrentContext, []);
@@ -48,6 +60,37 @@ const CheckIn = props => {
       }
     },
     [dispatch, context, goToErrorPage, token],
+  );
+
+  useEffect(
+    () => {
+      if (
+        !isDayOfDemographicsFlagsEnabled ||
+        demographicsFlagsSent ||
+        demographicsFlagsEmpty ||
+        getShouldSendDemographicsFlags(window)
+      )
+        return;
+      api.v2
+        .patchDayOfDemographicsData(demographicsData)
+        .then(resp => {
+          if (resp.data.error || resp.data.errors) {
+            throw new Error();
+          } else {
+            setDemographicsFlagsSent(true);
+          }
+        })
+        .catch(() => {});
+    },
+    [
+      demographicsData,
+      demographicsFlagsEmpty,
+      demographicsFlagsSent,
+      getShouldSendDemographicsFlags,
+      goToErrorPage,
+      isDayOfDemographicsFlagsEnabled,
+      setDemographicsFlagsSent,
+    ],
   );
 
   if (!appointment) {
