+++ src/applications/check-in/hooks/useSessionStorage.jsx
@@ -81,18 +81,24 @@ const useSessionStorage = (isPreCheckIn = true, maxValidateAttempts = 3) => {
     };
   };
 
-  const setDemographicsConfirmed = useCallback(
+  const setShouldSendDemographicsFlags = useCallback(
     (window, value) => {
-      setSessionKey(window, SESSION_STORAGE_KEYS.DEMOGRAPHICS_CONFIRMED, value);
+      setSessionKey(
+        window,
+        SESSION_STORAGE_KEYS.SHOULD_SEND_DEMOGRAPHICS_FLAGS,
+        value,
+      );
     },
     [SESSION_STORAGE_KEYS],
   );
 
-  const getDemographicsConfirmed = useCallback(
+  const getShouldSendDemographicsFlags = useCallback(
     window => {
       return (
-        getSessionKey(window, SESSION_STORAGE_KEYS.DEMOGRAPHICS_CONFIRMED) ??
-        false
+        getSessionKey(
+          window,
+          SESSION_STORAGE_KEYS.SHOULD_SEND_DEMOGRAPHICS_FLAGS,
+        ) ?? false
       );
     },
     [SESSION_STORAGE_KEYS],
@@ -106,8 +112,8 @@ const useSessionStorage = (isPreCheckIn = true, maxValidateAttempts = 3) => {
     getPreCheckinComplete,
     incrementValidateAttempts,
     getValidateAttempts,
-    setDemographicsConfirmed,
-    getDemographicsConfirmed,
+    setShouldSendDemographicsFlags,
+    getShouldSendDemographicsFlags,
   };
 };
