+++ src/applications/check-in/utils/session-storage/index.js
@@ -6,7 +6,7 @@ const createSessionStorageKeys = ({ isPreCheckIn = true }) => {
     CURRENT_UUID: `${namespace}.current.uuid`,
     VALIDATE_ATTEMPTS: `${namespace}.validate.attempts`,
     COMPLETE: `${namespace}.complete`,
-    DEMOGRAPHICS_CONFIRMED: `${namespace}.demographics.confirmed`,
+    SHOULD_SEND_DEMOGRAPHICS_FLAGS: `${namespace}.should.send.demographics.flags`,
   };
 };
